requirements.txt
alembic~=1.10
psycopg2~=2.9
fastapi~=0.76
python-dotenv~=0.21
uvicorn~=0.20
SQLAlchemy[asyncio]~=1.4
asyncpg~=0.25
redis~=4.3
strawberry-graphql==0.149.2
pydantic[email]~=1.10
black
cryptography~=37.0
passlib[bcrypt]~=1.7
python-jose[cryptography]~=3.2
google-auth~=1.35
google-auth-oauthlib~=0.5
google-auth-httplib2~=0.1
python-multipart==0.0.5